requirements.txt
weaviate-client>=3.24.1
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
openai>=1.3.0
requests>=2.31.0